UseCaseとRepositoryをcommonにまとめる
