+++ app/src/main/java/com/github/kitakkun/foos/di/RepositoryModule.kt
@@ -1,16 +1,9 @@
 package com.github.kitakkun.foos.di
 
-import com.github.kitakkun.foos.common.repository.PostsRepository
-import com.github.kitakkun.foos.common.repository.PostsRepositoryImpl
-import com.github.kitakkun.foos.common.repository.ReactionsRepository
-import com.github.kitakkun.foos.common.repository.ReactionsRepositoryImpl
+import com.github.kitakkun.foos.common.repository.*
 import com.github.kitakkun.foos.di.FirebaseModule.provideFireStoreInstance
 import com.github.kitakkun.foos.di.FirebaseModule.provideFirebaseAuthInstance
 import com.github.kitakkun.foos.di.FirebaseModule.provideFirebaseStorageInstance
-import com.github.kitakkun.foos.user.repository.FollowRepository
-import com.github.kitakkun.foos.user.repository.FollowRepositoryImpl
-import com.github.kitakkun.foos.user.repository.UsersRepository
-import com.github.kitakkun.foos.user.repository.UsersRepositoryImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
