+++ app/src/main/java/com/github/kitakkun/foos/ui/view/screen/home/HomeViewModelImpl.kt
@@ -8,8 +8,8 @@ import androidx.lifecycle.viewModelScope
 import com.github.kitakkun.foos.common.navigation.BottomSheet
 import com.github.kitakkun.foos.common.navigation.StringList
 import com.github.kitakkun.foos.common.navigation.SubScreen
+import com.github.kitakkun.foos.common.usecase.FetchPostsUseCase
 import com.github.kitakkun.foos.customview.composable.post.PostItemUiState
-import com.github.kitakkun.foos.data.domain.fetcher.post.FetchPostsUseCase
 import com.google.gson.Gson
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
