+++ app/src/main/java/com/github/kitakkun/foos/ui/view/screen/postdetail/PostDetailViewModelImpl.kt
@@ -6,8 +6,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.kitakkun.foos.common.model.DatabaseReaction
 import com.github.kitakkun.foos.common.repository.ReactionsRepository
+import com.github.kitakkun.foos.common.usecase.FetchPostByPostIdUseCase
 import com.github.kitakkun.foos.customview.composable.post.PostItemUiState
-import com.github.kitakkun.foos.data.domain.fetcher.post.FetchPostByPostIdUseCase
 import com.github.kitakkun.foos.post.postdetail.PostDetailScreenUiState
 import com.github.kitakkun.foos.post.postdetail.PostDetailViewModel
 import com.google.firebase.auth.ktx.auth
