+++ app/src/main/java/com/github/kitakkun/foos/ui/view/screen/reaction/ReactionViewModelImpl.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.github.kitakkun.foos.data.domain.FetchReactionsByUserIdUseCase
+import com.github.kitakkun.foos.common.usecase.FetchReactionsByUserIdUseCase
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 import dagger.hilt.android.lifecycle.HiltViewModel
