+++ common/src/main/java/com/github/kitakkun/foos/common/model/DatabaseFollow.kt
@@ -1,4 +1,4 @@
-package com.github.kitakkun.foos.user.model
+package com.github.kitakkun.foos.common.model
 
 import com.google.firebase.firestore.ServerTimestamp
 import java.util.*
