+++ common/src/main/java/com/github/kitakkun/foos/common/repository/FollowRepositoryImpl.kt
@@ -1,7 +1,7 @@
-package com.github.kitakkun.foos.user.repository
+package com.github.kitakkun.foos.common.repository
 
 import com.github.kitakkun.foos.common.ext.join
-import com.github.kitakkun.foos.user.model.DatabaseFollow
+import com.github.kitakkun.foos.common.model.DatabaseFollow
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FieldValue
 import com.google.firebase.firestore.FirebaseFirestore
