+++ common/src/main/java/com/github/kitakkun/foos/common/repository/UsersRepository.kt
@@ -1,12 +1,16 @@
-package com.github.kitakkun.foos.user.repository
+package com.github.kitakkun.foos.common.repository
 
 import com.github.kitakkun.foos.common.model.DatabaseUser
-import com.github.kitakkun.foos.user.model.Email
-import com.github.kitakkun.foos.user.model.Password
+import com.github.kitakkun.foos.common.model.Email
+import com.github.kitakkun.foos.common.model.Password
 import com.github.michaelbull.result.Result
 
 interface UsersRepository {
-    suspend fun create(email: Email, password: Password): Result<DatabaseUser, Throwable>
+    suspend fun create(
+        email: Email,
+        password: Password
+    ): Result<DatabaseUser, Throwable>
+
     suspend fun create(databaseUser: DatabaseUser)
     suspend fun fetchByUserId(userId: String): DatabaseUser?
     suspend fun fetchByUserIds(userIds: List<String>): List<DatabaseUser>
