+++ common/src/main/java/com/github/kitakkun/foos/common/usecase/FetchPostByDatabasePostUseCase.kt
@@ -1,9 +1,9 @@
-package com.github.kitakkun.foos.data.domain.fetcher.post
+package com.github.kitakkun.foos.common.usecase
 
 import com.github.kitakkun.foos.common.model.DatabasePost
 import com.github.kitakkun.foos.common.model.Post
 import com.github.kitakkun.foos.common.repository.ReactionsRepository
-import com.github.kitakkun.foos.user.repository.UsersRepository
+import com.github.kitakkun.foos.common.repository.UsersRepository
 import javax.inject.Inject
 
 /**
