+++ common/src/main/java/com/github/kitakkun/foos/common/usecase/FetchPostsByUserIdUseCase.kt
@@ -1,39 +1,41 @@
-package com.github.kitakkun.foos.data.domain.fetcher.post
+package com.github.kitakkun.foos.common.usecase
 
 import com.github.kitakkun.foos.common.model.Post
 import com.github.kitakkun.foos.common.repository.PostsRepository
 import com.github.kitakkun.foos.common.repository.PostsRepositoryImpl
 import com.github.kitakkun.foos.common.repository.ReactionsRepository
-import com.github.kitakkun.foos.user.repository.UsersRepository
+import com.github.kitakkun.foos.common.repository.UsersRepository
 import java.util.*
 import javax.inject.Inject
 
 /**
- * フィルタなしで投稿をフェッチするユースケース
+ * ユーザIDを元に投稿をフェッチするユースケース
  */
-class FetchPostsUseCase @Inject constructor(
+class FetchPostsByUserIdUseCase @Inject constructor(
     private val postsRepository: PostsRepository,
     private val usersRepository: UsersRepository,
     private val reactionsRepository: ReactionsRepository,
 ) {
 
-    suspend operator fun invoke(start: Date? = null, end: Date? = null): List<Post> {
-        val dbPosts = postsRepository.fetch(start, end, PostsRepositoryImpl.DEFAULT_LOAD_LIMIT)
-        val dbUsers = usersRepository.fetchByUserIds(dbPosts.map { it.userId })
-        val dbReactions = reactionsRepository.fetchByPostIds(dbPosts.map { it.postId })
+    suspend operator fun invoke(userId: String, end: Date = Date()): List<Post> {
+        val dbUser = usersRepository.fetchByUserId(userId)
+        val dbPosts = postsRepository.fetchByUserId(
+            userId,
+            null,
+            end,
+            PostsRepositoryImpl.DEFAULT_LOAD_LIMIT
+        )
+        val dbReactions = dbPosts.map { it.postId }.let { reactionsRepository.fetchByPostIds(it) }
 
         return dbPosts.mapNotNull { post ->
-            val user = dbUsers.find { it.userId == post.userId }
             val reactions = dbReactions.filter { it.postId == post.postId }
-            user?.let {
+            dbUser?.let {
                 Post(
                     post = post,
-                    user = user,
+                    user = dbUser,
                     reaction = reactions,
                 )
             }
         }
-
     }
-
 }
